fix: s:WARNING is not defined in the os module
If we don't want the variables to be globals, we must use a helper
function to return the correct value from the netrw.vim autoload script.

Also, it seems that when netrw#os#Execute() fails, this should indicate
an ERROR and not just a WARNING, so let's just increase the loglevel for
the error case.

Signed-off-by: Christian Brabandt <cb@256bit.org>